MLPO/requirements.txt
pandas
numpy
matplotlib
seaborn
yfinance
statsmodels
scikit-learn
plotly
xgboost
scipy
streamlit
python-dateutil
pytest
